+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.1 [In Progess]
 * [ADD] Add new icons
+* [IMP] Optimisation for TextDrawing font
 
 ## 1.0.0
 * [ADD] Update Bluetooth communication for aRdent 1